chore: moves test utils
