--- bot/src/__test__/__test_util__/buttons/index.ts
@@ -1,13 +1,0 @@
-import { KeyboardButton } from "node-telegram-bot-api";
-
-export function generateTelegramButtons(
-  buttonsQty: number
-): Array<KeyboardButton> {
-  const telegramButtons = Array(buttonsQty)
-    .fill({ text: "test" })
-    .map((button, i) => {
-      return { text: `${button.text} ${i}` };
-    });
-
-  return telegramButtons;
-}
