--- bot/src/__test__/__test_util__/index.ts
@@ -1,1 +1,0 @@
-export * from "./buttons";
