+++ bot/src/commands/index.ts
@@ -1,3 +1,1 @@
-export * from "./CreateWalletCommand";
 export * from "./StartCommand";
-export * from "./TopUpCommand";
