+++ bot/src/commands/wallet/index.ts
@@ -1,3 +1,5 @@
 export * from "./BalanceCommand";
+export * from "./CreateWalletCommand";
 export * from "./SetCountryCodeCommand";
 export * from "./SetCurrencyCodeCommand";
+export * from "./TopUpCommand";
